backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
psycopg2-binary>=2.9.11
sqlalchemy>=2.0.23
python-dotenv>=1.0.0
openpyxl>=3.1.0
python-multipart>=0.0.6
pydantic>=2.5.0
gunicorn>=21.0.0
